TRUS-2494 Refactoring trust store connector
